requirements.txt
Django==1.9.7
PyJWT==1.4.0
argparse==1.2.1
django-adminplus==0.5
django-autocomplete-light==3.1.6
django-bootstrap3==7.0.1
django-ckeditor==5.0.3
django-tinymce==2.3.0
gunicorn==19.6.0
oauthlib==1.1.2
psycopg2==2.6.1
python-openid==2.2.5
python-social-auth==0.2.19
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
#wsgiref==0.1.2
